+++ fiddles/angular2-cli/fiddle-0002-GeneratePipe/e2e/app.po.ts
@@ -1,0 +1,9 @@
+export class GeneratePipePage {
+  navigateTo() {
+    return browser.get('/');
+  }
+
+  getParagraphText() {
+    return element(by.css('generate-pipe-app h1')).getText();
+  }
+}