+++ fiddles/angular2-cli/fiddle-0002-GeneratePipe/src/app/generate-pipe.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  selector: 'generate-pipe-app',
+  templateUrl: 'generate-pipe.component.html',
+  styleUrls: ['generate-pipe.component.css']
+})
+export class GeneratePipeAppComponent {
+  title = 'generate-pipe works!';
+}