+++ fiddles/angular2-cli/fiddle-0002-GeneratePipe/src/app/index.ts
@@ -1,0 +1,2 @@
+export {environment} from './environment';
+export {GeneratePipeAppComponent} from './generate-pipe.component';