+++ fiddles/angular2-cli/fiddle-0002-GeneratePipe/src/app/shared/grid/grid.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  selector: 'app-grid',
+  templateUrl: 'grid.component.html',
+  styleUrls: ['grid.component.css']
+})
+export class GridComponent implements OnInit {
+
+  constructor() {}
+
+  ngOnInit() {
+  }
+
+}