+++ fiddles/angular2/fiddle-0039-ResponsiveModalPopUp/app/app.component.ts
@@ -1,0 +1,17 @@
+import {Component} from '@angular/core';
+import {SidebarComponent} from './sidebar/sidebar.component';
+import {ContentComponent} from './content/content.component';
+import {GlobalService} from './global/global.service';
+
+
+@Component({
+    selector: 'app',
+    templateUrl: './app/app.component.html',
+    styleUrls: ['./app/app.component.css'],
+    directives: [[SidebarComponent, ContentComponent]],
+    providers: [GlobalService]
+})
+export class AppComponent {
+
+
+}