+++ fiddles/angular2/fiddle-0039-ResponsiveModalPopUp/app/content/content-popup/content-popup.component.spec.ts
@@ -1,0 +1,46 @@
+import {
+  beforeEach,
+  beforeEachProviders,
+  describe,
+  expect,
+  it,
+  inject,
+} from '@angular/core/testing';
+import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
+import { Component } from '@angular/core';
+import { By } from '@angular/platform-browser';
+import { ContentPopupComponent } from './content-popup.component';
+
+describe('Component: ContentPopup', () => {
+  let builder: TestComponentBuilder;
+
+  beforeEachProviders(() => [ContentPopupComponent]);
+  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
+    builder = tcb;
+  }));
+
+  it('should inject the component', inject([ContentPopupComponent],
+      (component: ContentPopupComponent) => {
+    expect(component).toBeTruthy();
+  }));
+
+  it('should create the component', inject([], () => {
+    return builder.createAsync(ContentPopupComponentTestController)
+      .then((fixture: ComponentFixture<any>) => {
+        let query = fixture.debugElement.query(By.directive(ContentPopupComponent));
+        expect(query).toBeTruthy();
+        expect(query.componentInstance).toBeTruthy();
+      });
+  }));
+});
+
+@Component({
+  selector: 'test',
+  template: `
+    <content-popup></content-popup>
+  `,
+  directives: [ContentPopupComponent]
+})
+class ContentPopupComponentTestController {
+}
+