+++ fiddles/angular2/fiddle-0039-ResponsiveModalPopUp/app/content/content-popup/index.ts
@@ -1,0 +1,1 @@
+export { ContentPopupComponent } from './content-popup.component';