+++ fiddles/angular2/fiddle-0039-ResponsiveModalPopUp/app/meta.ts
@@ -1,0 +1,7 @@
+var metadata = {
+    fiddleHeader: 'Angular 2 - Modal PopUp',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0039-ResponsiveModalPopUp'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};